modify Lab4

+++ src/MyJPanel.java
@@ -9,8 +9,8 @@ import javax.swing.*;
 public class MyJPanel extends JPanel implements MouseMotionListener{
 	
 	private Square square = new Square(100);
-	private float centerX;
-	private float centerY;
+	private int centerX;
+	private int centerY;
 	public MyJPanel(){
 		addMouseMotionListener(this);
 	}
